Split grpcutil and errutil into client/server pkgs
This way, code in the client can only depend on other code in src/client
