+++ src/server/pkg/collection/range.go
@@ -7,7 +7,7 @@ import (
 
 	etcd "github.com/coreos/etcd/clientv3"
 	"github.com/coreos/etcd/mvcc/mvccpb"
-	"github.com/pachyderm/pachyderm/src/server/pkg/errutil"
+	"github.com/pachyderm/pachyderm/src/client/pkg/errutil"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
